refactor(expense): reorganize expense hooks and update imports for consistency
