+++ apps/web/@component/expense/expenseForm.tsx
@@ -5,7 +5,7 @@ import {
   UpsertExpenseInput,
   useCreateExpense,
   useUpdateExpense,
-} from "./useExpense";
+} from "../../@hook/useExpense";
 import { isTodayOrPastYMD } from "../../@utils/date/IsTodayorPast";
 
 export function ExpenseForm({ selectedDate }: { selectedDate: string }) {
